Update Course.java
Added a competency data member to course to keep track of whether it fullfills a competency. It is default set to N which means no competency, when making list of courses make an if statement to check if there is a last element of whatever number. If there is then setCompetency of that value.