Update Schedule.java
Added lab course for the time being as some intro courses require a lab. Also began working on checkConflict. For the time being when testing we will assume there is no time conflicts. checkConflict is intended to compare times to see if there is a conflict, if there isn't any it returns true if there is it returns false.